requirements.txt
aiofiles==23.2.1
python-dotenv==1.0.0